v2: fix ProductionOrder float-as-string crash en generar imputaciones SAP
SapOrders.Order se almacenaba como "2000097595.0" (float stringificado)
y TablaCentral.ProductionOrder es BIGINT. Ahora se limpia con int(float())
al insertar, y el ETL usa Int64 para evitar el .0 en futuros uploads.

+++ backend/app/services/sap_etl_utils.py
@@ -21,7 +21,7 @@ def transformar_datos_sap(df: pd.DataFrame) -> pd.DataFrame:
     df['OperationActivityFull'] = df['Operation Activity']
     df['EffectivityFull'] = df['Effectivity']
     # Este parseo es un ejemplo del original:
-    df['Order'] = pd.to_numeric(df['Order'].str.extract(r'\((\d+)\)')[0], errors='coerce')
+    df['Order'] = pd.to_numeric(df['Order'].str.extract(r'\((\d+)\)')[0], errors='coerce').astype('Int64')
 
     df['Operation'] = df['Operation Activity'].str.extract(r'(\d+)')
     df['OperationActivity'] = df['Operation Activity'].str.extract(r'^(\S+)')
